feat(microadaptivity): make end feedback sticky
