+++ packages/editor/src/plugins/text-area-exercise/renderer.tsx
@@ -5,7 +5,6 @@ import { useState } from 'react'
 
 import { Blocks } from './blocks'
 import { PrototypeStateStore } from './prototype-state'
-import { usePluginId } from './use-plugin-id'
 import { useTextAreaPluginStateValues } from './use-text-area-plugin-state-values'
 
 // Schüly Ansicht
@@ -26,7 +25,7 @@ export function TextAreaExerciseRenderer({
   return (
     <>
       {!isInEditor && allowShowSolutionStrategy && solutionStrategy ? (
-        <div className="my-5 whitespace-pre-wrap rounded-xl bg-brand-100 p-5">
+        <div className="my-5 w-full whitespace-pre-wrap rounded-xl bg-brand-100 p-5">
           <div className="mb-3 font-bold">Remember</div>
           <div className="flex flex-col gap-2">
             {solutionStrategy.split('\n').map((elem, index) => {
